pyproject.toml
[build-system]
requires = ["setuptools>=61"]
build-backend = "setuptools.build_meta"

[project]
name = "comfyui-glm-image"
version = "0.1.0"
description = "ComfyUI nodes for GLM-Image model integration (loaders + samplers)"
readme = "README.md"
license = { text = "MIT" }
requires-python = ">=3.10"
authors = [{ name = "Likhith-24" }]
dependencies = [
    "torch>=2.0",
    "diffusers>=0.27",
    "transformers>=4.40",
    "pillow>=10.0",
    "huggingface_hub>=0.23",
    "accelerate",
]

[project.urls]
Repository = "https://github.com/Likhith-24/ComfyUI-GLM_Image"
Issues = "https://github.com/Likhith-24/ComfyUI-GLM_Image/issues"

[tool.comfy]
PublisherId = "likhith-24"
DisplayName = "ComfyUI-GLM_Image"
Icon = "https://i.pinimg.com/736x/5d/52/19/5d5219ef27c1530bea027ff95d352e05.jpg"
